updates to shared
